pyproject.toml
[tool.poetry]
name = "tdp-server"
version = "0.1.0"
description = ""
authors = ["Guillaume Boutry <diode-boutry.consultant@dgfip.finances.gouv.fr>"]

[tool.poetry.dependencies]
python = ">=3.6.2,<4.0"
fastapi = { version = "0.75.0" }
tdp-lib = { git = "https://github.com/TOSIT-IO/tdp-lib" }
alembic = "1.7.7"
tenacity = "8.0.1"
psycopg2-binary = { version ="2.9.5", optional = true}
Authlib = "1.0.0"
httpx = "0.22.0"
python-jose = { version = "3.3.0", extras = ["cryptography"] }
filelock = "3.4.1"
python-multipart = "^0.0.5"
itsdangerous = "2.0.1"

[tool.poetry.dev-dependencies]
pytest = "6.2.5"
requests = "2.27.1"
black = "21.12b0"
isort = "5.10.1"
uvicorn = { version = "0.16.0", extras = ["standard"] }
pre-commit = "2.17.0"

[tool.poetry.extras]
pg = ["psycopg2-binary"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ['py36']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.venv
    | venv
    | _build
    | build
    | dist
  )/
)
'''

[tool.isort]
src_paths = ["tdp_server"]
force_single_line = false
known_first_party = "tdp_server"
known_third_party = "tdp"
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88
multi_line_output = 3
